requirements.txt
aiogram==3.16.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
requests==2.32.3
apscheduler==3.11.0
aiogram-calendar==0.6.0
langchain==0.3.14
langchain-gigachat==0.3.3